Update ODC name in OdcHelper.cs (#1417)
When using Analyze in Excel, the connection name was set to "mtbsql608v-dev_mssqlinst01 PRS Model", which is confusing and sounds like the name of a local model from our appreciated DaxStudio devs. Suggesting to change it to "DAX Studio Analyze in Excel PowerBI Model" to make it easier to understand. Thanks!